commit2

+++ tambah_config.php
@@ -16,10 +16,10 @@ function CreateData()
     $nama = mysqli_real_escape_string($GLOBALS['link'], trim($_POST['nama_diklat']));
     $jenis = mysqli_real_escape_string($GLOBALS['link'], trim($_POST['jenis_diklat']));
     $jalan = mysqli_real_escape_string($GLOBALS['link'], trim($_POST['penyelenggara_diklat']));
-    $jumlah = (int)mysqli_real_escape_string($GLOBALS['link'], trim($_POST['jumlah_peserta']));
-    $mulai = (int)mysqli_real_escape_string($GLOBALS['link'], trim($_POST['mulai']));
+    $jumlah = mysqli_real_escape_string($GLOBALS['link'], trim($_POST['jumlah_peserta']));
+    $mulai = mysqli_real_escape_string($GLOBALS['link'], trim($_POST['mulai']));
     $selesai = mysqli_real_escape_string($GLOBALS['link'], trim($_POST['selesai']));
-    $durasi = (int)mysqli_real_escape_string($GLOBALS['link'], trim($_POST['durasi']));
+    $durasi = mysqli_real_escape_string($GLOBALS['link'], trim($_POST['durasi']));
     $lokasi = mysqli_real_escape_string($GLOBALS['link'], trim($_POST['lokasi']));
 
     if ($nama && $jenis && $jalan && $jumlah && $mulai && $selesai && $durasi && $lokasi) {
